requirements.txt
pydub
tqdm
Pillow